op-batcher: Restructure source code
This moves everything into a `batcher` directory. It's maybe not
the perfect name yet, but this is getting split into enough components
that having all of the source files in the top level directory was
becoming annoying.
