+++ op-batcher/batcher/driver.go
@@ -1,4 +1,4 @@
-package op_batcher
+package batcher
 
 import (
 	"context"
@@ -10,7 +10,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/ethereum-optimism/optimism/op-batcher/sequencer"
 	"github.com/ethereum-optimism/optimism/op-node/eth"
 	"github.com/ethereum-optimism/optimism/op-proposer/txmgr"
 	"github.com/ethereum/go-ethereum/accounts"
@@ -25,7 +24,7 @@ import (
 type BatchSubmitter struct {
 	txMgr txmgr.TxManager
 	addr  common.Address
-	cfg   sequencer.Config
+	cfg   Config
 	wg    sync.WaitGroup
 	done  chan struct{}
 	log   log.Logger
@@ -40,7 +39,7 @@ type BatchSubmitter struct {
 
 // NewBatchSubmitter initializes the BatchSubmitter, gathering any resources
 // that will be needed during operation.
-func NewBatchSubmitter(cfg Config, l log.Logger) (*BatchSubmitter, error) {
+func NewBatchSubmitter(cfg CLIConfig, l log.Logger) (*BatchSubmitter, error) {
 	ctx := context.Background()
 
 	var err error
@@ -125,7 +124,7 @@ func NewBatchSubmitter(cfg Config, l log.Logger) (*BatchSubmitter, error) {
 		SafeAbortNonceTooLowCount: cfg.SafeAbortNonceTooLowCount,
 	}
 
-	batcherCfg := sequencer.Config{
+	batcherCfg := Config{
 		Log:               l,
 		Name:              "Batch Submitter",
 		L1Client:          l1Client,
